OCR BASICO + CARGA DE IMAGEN

+++ app/src/main/java/com/example/msc/domain/model/Products.kt
@@ -4,7 +4,7 @@ package com.example.msc.domain.model
 data class Products(
     val name : String = "",
     val price : Double = 0.0,
-    val quantity : Int = 0
+    val quantity : Double = 0.0
 ){
     fun total() : Double {
         return price * quantity
